feat: Add createId and zValidator to accounts API
This commit adds the `createId` function from the `@paralleldrive/cuid2` library and the `zValidator` function from the `@hono/zod-validator` library to the accounts API in the `accounts.ts` file. It also updates the API endpoint to include a POST request that validates and inserts a new account into the database.